+++ crates/scouter/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "scouter"
+version = { workspace = true }
+edition = { workspace = true }
+repository = { workspace = true }
+
+authors = [
+  "Steven Forrester <sjforrester32@gmail.com>"
+]
+
+license = "MIT"
+description = "Rust Scouter logic"
+
+[dependencies]
+pyo3 = { workspace = true }
+serde = { version = "1.0.163", features = ["derive"] }
+itertools = "0.11.0"
+rayon = "1.7.0"
+noisy_float = "0.2.0"
+anyhow = "1.0.79"
+chrono = "0.4.38"
+num-traits = "0.2.19"
+serde_json = "1.0.117"
+indicatif = "0.17.8"
+ndarray = { version = "0.15.6", features = ["rayon"] }
+numpy = { workspace = true }
+ndarray-stats = "0.5.1"
+
+[dev-dependencies]
+approx = "0.5.1"
+ndarray-rand = "0.14.0"
+criterion = { version = "0.4", features = ["html_reports"] }
+
+[profile.release]
+lto = "fat"
+codegen-units = 1
+strip = true
