+++ crates/scouter/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod math;
+pub mod types;
