+++ crates/scouter/src/math/mod.rs
