+++ crates/scouter/src/math/monitor.rs
