+++ crates/scouter/src/math/profiler.rs
