+++ py-scouter/Cargo.toml
@@ -1,0 +1,20 @@
+[package]
+name = "py-scouter"
+version = "0.1.0-rc.1"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+name = "scouter"
+crate-type = ["cdylib", "rlib"]
+
+
+[dependencies]
+scouter = { workspace = true }
+pyo3 = { version = "0.21", features = ["extension-module"] }
+numpy = "0.21.0"
+
+[profile.release]
+lto = "fat"
+codegen-units = 1
+strip = true
