+++ py-scouter/scouter/_scouter.pyi
