+++ py-scouter/scouter/utils/comparison/numpy_scouter.py
